pyproject.toml
[project]
name = "ai-paper-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "bs4>=0.0.2",
    "click>=8.1.8",
    "llama-index-core>=0.12.25",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-llms-openai>=0.3.25",
    "llama-index-utils-workflow>=0.3.0",
    "requests>=2.32.3",
    "scholarly>=1.7.11",
    "slugify>=0.0.1",
]
